refactored board classes
